Never give up on a module.
We've observed gardener "giving up" on a malfunctioning service in
production; this essentially guarantees that the module will stop
functioning. If gardener were to keep trying, there's at least some
possibility that portions of that module will keep working, e.g. along
different code paths, or when the module is provided with input that
doesn't trigger the crash/exit.

The `spinSleepTime` option for `forever` is poorly documented, but at
least one string in the codebase leads me to believe that `forever` will
give up on a service if `minUptime` isn't reached *and* `spinSleepTime`
is undefined:

  https://github.com/nodejitsu/forever/blob/master/lib/forever/cli.js#L229

This patch simply sets `spinSleepTime`, so `forever` (and thus,
`gardener`) will wait for two seconds and retry rather than give up on
the module entirely.