added console error when couchdb is not found.
The logger can't flush the log on process.exit unless we wait for the
stream event, switched to using console.log for sanity right now.  This
error is fairly uncommon (initial couchdb bind being unavailable) so
having the console log is an improvement from only having an exit code
of 1 to decipher the problem.

A logger refactor is probably in order to confirm logging succeeds when
processes exit.

More info: https://github.com/winstonjs/winston/issues/228